rename utils checker file and add helper functions to it
